+++ ui.py
@@ -32,8 +32,7 @@ class MGB_PT_sidebar(Panel):
             return
 
         col = layout.column(align=True)
-        col.prop(p, "enable_mmb_transform", text="MMB Transform")
-        col.prop(p, "remap_orbit_to_rmb", text="RMB Orbit")
+        col.prop(p, "enable_rmb_transform", text="RMB Transform")
 
         layout.operator("mgb.reset_keymap", icon='LOOP_BACK')